Move setup_project logic to a separate file.
