+++ crates/compiler/Cargo.toml
@@ -16,6 +16,7 @@ syntax = { path = "../syntax" }
 parser = { path = "../parser" }
 project = { path = "../project" }
 salsa.workspace = true
+thiserror.workspace = true
 
 [[bin]]
 name = "cairo-compile"
