+++ crates/compiler/src/cli.rs
@@ -1,22 +1,19 @@
-use std::ffi::OsStr;
 use std::fs;
 use std::path::Path;
 use std::process::ExitCode;
 use std::sync::Arc;
 
 use clap::Parser;
+use compiler::db::RootDatabase;
+use compiler::project::setup_project;
 use defs::db::DefsGroup;
 use defs::ids::ModuleId;
-use filesystem::db::{AsFilesGroupMut, FilesGroup, FilesGroupEx};
-use filesystem::ids::{CrateLongId, Directory, FileLongId};
+use filesystem::db::FilesGroup;
+use filesystem::ids::FileLongId;
 use lowering::db::LoweringGroup;
-use project::ProjectConfig;
+use parser::db::ParserGroup;
 use semantic::db::SemanticGroup;
 use sierra_generator::db::SierraGenGroup;
-
-mod db;
-use db::RootDatabase;
-use parser::db::ParserGroup;
 use sierra_generator::replace_ids::replace_sierra_ids_in_program;
 
 /// Command line args parser.
@@ -63,52 +60,14 @@ fn print_diagnostics(db: &mut RootDatabase) {
     }
 }
 
-fn setup_single_file_crate(db: &mut RootDatabase, path: &Path) -> Result<(), ()> {
-    match path.extension().and_then(OsStr::to_str) {
-        Some("cairo") => (),
-        _ => {
-            eprintln!("Only files with .cairo extension can be compiled.");
-            return Err(());
-        }
-    }
-
-    if let Some(file_stemp) = path.file_stem().and_then(OsStr::to_str) {
-        if file_stemp != "lib" {
-            // If file_stemp is not lib, create a fake lib file.
-
-            if !path.exists() {
-                eprintln!("Couldn't read {}: No such file", path.to_string_lossy());
-                return Err(());
-            }
-
-            let crate_id = db.intern_crate(CrateLongId(file_stemp.into()));
-            db.set_crate_root(crate_id, Some(Directory(path.parent().unwrap().to_path_buf())));
-
-            let module_id = ModuleId::CrateRoot(crate_id);
-            let file_id = db.module_file(module_id).unwrap();
-            db.as_files_group_mut()
-                .override_file_content(file_id, Some(Arc::new(format!("mod {};", file_stemp))));
-        }
-    }
-    Ok(())
-}
-
 fn main() -> ExitCode {
     let args = Args::parse();
 
     let mut db_val = RootDatabase::default();
     let db = &mut db_val;
 
-    let path = Path::new(&args.path);
-    if path.is_dir() {
-        match ProjectConfig::from_directory(path) {
-            Ok(config) => db.with_project_config(config),
-            _ => {
-                eprintln!("Failed to load project config.");
-                return ExitCode::FAILURE;
-            }
-        };
-    } else if setup_single_file_crate(db, path).is_err() {
+    if let Err(error) = setup_project(db, Path::new(&args.path)) {
+        eprintln!("{}", error);
         return ExitCode::FAILURE;
     }
